+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 - Target class in field decorator instead of prototype
+- Append Core options to handle index prefix
+- Update tests by mocking when relevant
 
 ## [1.0.0] 2018-04-21